Remove .config/fcitx/
It would be nice to include .config/fcitx/. However currently, the
directory is cluttered and contain file that change over time even
without changing the configuration such as database and history which
should not be in the directory in the first place.

It would be a pain to maintain such a directory with git.