Yet another example of *smart* software
For context, I am using ssh support from GnuPG. Whenever we try to
connect to the ssh agent for example during a git push, we will connect
to GnuPG instead of something like OpenSSH. GnuPG will in turn ask for a
master password by invoking pinentry. This will dispatch to pinentry-gtk
in my case because I am running in a graphical environment.

The problem is that I need to run: $ gpg-connect-agent updatestartuptty
/bye before attempting to use the ssh support or else pinentry will
fail. This is perplexing to me because according to the documentation of
UPDATESTARTUPTTY, it is responsible for setting up the start TTY and
X-DISPLAY variables of the session. However, 1: pinentry-gtk obviously
does not run on a tty, 2: I am running wayland so DISPLAY is irrelevant.

It turns out in GnuPG source code, it try to detect if DISPLAY variable
is available and fallback to the implementation in pinentry-tty.

Hence, I need to somehow inject DISPLAY variable into gpg-agent even
though it is completely unnecessary because gtk run perfectly fine on
wayland.

Because gpg-agent actually launch via systemd socket activation
**sigh**, this is done by using dbus-update-activation-environment.

TL;DR:
 - pinentry thinks it can launch pinentry-gtk because WAYLAND_DISPLAY is
   set
 - pinentry-gtk disagree because DISPLAY is not set and fallback to
   pinentry-tty
 - updatestartuptty command set DISPLAY and make pinentry and
   pinentry-gtk agree with each other